Add large numbers and long decimals to be in scientific notation
Previously, large numbers and long decimals were not being rounded correctly. In the case of large numbers, the calculation was way off because the ends were chopped to fit the display. On the other hand, long decimals along with scientific notation were incorrectly rounded and ended up with larger numbers than the true calculation. This was fixed by converting both types to scientific notation.